+++ travel-planner/src/interfaces/CityActivities.ts
@@ -1,0 +1,14 @@
+// src/api/cityActivities.ts
+export interface Activity {
+  id: string;
+  name: string;
+  shortDescription?: string;
+  description: string;
+  price: {
+    amount: string;
+    currencyCode: string;
+  };
+  pictures: string[];
+  bookingLink: string;
+  minimumDuration: string;
+}